requirements.txt
Django
psycopg2
djangorestframework==3.11.0,<3.12.0
django-tinymce
pillow==7.0.0
gunicorn
